+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ticketsoup",
-  "version": "0.5.0-alpha",
+  "version": "0.5.0",
   "private": true,
   "type": "module",
   "scripts": {
